rename views->layouts folder
